Removed unused or unneeded telemetry-related code, and obsoleted the file joint_support.h
